AST: Setup AST, helper functions, and inits for each file

+++ src/token/types.h
@@ -131,8 +131,8 @@ typedef struct {
     Arena arena;
 
     Token* items;
-    u32 count;
-    u32 capacity;
+    u64 count;
+    u64 capacity;
 } TokenArray;
 
 #endif // !LILY_TOKEN_TYPES_H
